+++ src/main/java/pl/kedrabartosz/HomeBudget/version1/Person.java
@@ -1,4 +1,4 @@
-package pl.kedrabartosz.HomeBudget;
+package pl.kedrabartosz.HomeBudget.version1;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
