+++ src/main/java/pl/kedrabartosz/HomeBudget/version1/repository/FileBasedCategoryRepository.java
@@ -1,6 +1,6 @@
-package pl.kedrabartosz.HomeBudget.repository;
+package pl.kedrabartosz.HomeBudget.version1.repository;
 
-import pl.kedrabartosz.HomeBudget.Category;
+import pl.kedrabartosz.HomeBudget.version1.Category;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
