+++ src/main/java/pl/kedrabartosz/HomeBudget/version1/repository/ItemRepository.java
@@ -1,7 +1,7 @@
-package pl.kedrabartosz.HomeBudget.repository;
+package pl.kedrabartosz.HomeBudget.version1.repository;
 
-import pl.kedrabartosz.HomeBudget.Category;
-import pl.kedrabartosz.HomeBudget.Item;
+import pl.kedrabartosz.HomeBudget.version1.Category;
+import pl.kedrabartosz.HomeBudget.version1.Item;
 
 import java.util.List;
 import java.util.Optional;
