+++ src/main/java/pl/kedrabartosz/HomeBudget/version1/repository/ListBasedReceiptRepository.java
@@ -1,9 +1,9 @@
-package pl.kedrabartosz.HomeBudget.repository;
+package pl.kedrabartosz.HomeBudget.version1.repository;
 
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Repository;
-import pl.kedrabartosz.HomeBudget.Person;
-import pl.kedrabartosz.HomeBudget.Receipt;
+import pl.kedrabartosz.HomeBudget.version1.Person;
+import pl.kedrabartosz.HomeBudget.version1.Receipt;
 
 import java.util.ArrayList;
 import java.util.List;
