+++ src/main/java/pl/kedrabartosz/HomeBudget/version1/service/CategoryService.java
@@ -1,11 +1,11 @@
-package pl.kedrabartosz.HomeBudget.service;
+package pl.kedrabartosz.HomeBudget.version1.service;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import pl.kedrabartosz.HomeBudget.Category;
-import pl.kedrabartosz.HomeBudget.repository.CategoryRepository;
+import pl.kedrabartosz.HomeBudget.version1.Category;
+import pl.kedrabartosz.HomeBudget.version1.repository.CategoryRepository;
 
 @Service
 //service sie dodaje dla springa! nie jest to koniecznie bez tego tez by smigalo :D!
